Tag cloud ☁️  added to ag d8-1803 - Added service to embed node_add_tags view - Reworked other tag clouds to use new service
