+++ modules/access_misc/access_misc.module
@@ -11,7 +11,6 @@ use Drupal\Core\Entity\EntityInterface;
 use Drupal\views\ViewExecutable;
 use Drupal\taxonomy\Entity\Term;
 use Drupal\access_misc\Plugin\Util\EventDateConvert;
-use Drupal\views\Views;
 
 /**
  * Implements hook_views_post_render().
@@ -109,17 +108,11 @@ function access_misc_datalayer_alter(&$data_layer) {
  * Implements hook_form_alter().
  */
 function access_misc_form_alter(&$form, FormStateInterface $form_state, $form_id) {
-  if ($form_id == 'node_affinity_group_edit_form' || $form_id == 'node_affinity_group_form') {
-    $form['#attached']['library'][] = 'access_misc/add_select';
-  }
   // Custom tags on /form/ci-link webform.
   if ($form_id == 'webform_submission_resource_add_form' || $form_id == 'webform_submission_resource_edit_form') {
     $form['#attached']['library'][] = 'access_misc/custom_tags';
-    $view = Views::getView('node_add_tags');
-    $view->setDisplay('block_1');
-    $view->execute();
-    $rendered = $view->render();
-    $output = \Drupal::service('renderer')->render($rendered);
+    $add_tags = \Drupal::service('access_misc.addtags');
+    $output = $add_tags->getView();
     $form['custom_tags'] = [
       '#markup' => "<details open class='tags'><summary><label class='js-form-required form-required'>Select relevant tags</label></summary>$output</details>",
       '#weight' => 3,
