+++ modules/access_misc/css/node_add_tags.css
@@ -1,0 +1,20 @@
+#edit-field-tags-wrapper {
+  display: none;
+}
+
+details.tags summary {
+  cursor: pointer;
+}
+
+.tags-select {
+  display: inline-block;
+  border: 1px solid #000;
+  cursor: pointer;
+  padding: 5px;
+  margin: 5px;
+}
+
+.selected {
+  background-color: #1A5B6E;
+  color: #fff;
+}
